skills/paperbanana-figures/scripts/requirements.txt
google-genai>=1.0.0
openai>=1.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
pandas>=2.0.0
